backend.backup/requirements.txt
Django
djangorestframework
django-debug-toolbar
psycopg2
djangorestframework_simplejwt